[add] update build-iso.sh-iso to return 0 if the iso already exist

+++ tools/build-iso.sh
@@ -9,6 +9,13 @@
 ## $2 release name
 ## $3 version
 
+if test -f "$1.iso"; then
+  exit 0
+fi
+
+rm -rf sysroot/boot/
+mkdir -p sysroot/boot/grub
 cp "$1" sysroot/boot/"$1"
+./tools/generate-grub-config.sh $1 $2 $3 > sysroot/boot/grub/grub.cfg
 grub-mkrescue -o "$1.iso" sysroot
-exec ./tools/generate-grub-config.sh $1 $2 $3 > sysroot/boot/grub/grub.cfg
+
